+++ tvdcn/csrc/ops/cpu/deform_conv2d_kernels_cpu.cpp
@@ -188,28 +188,28 @@ namespace tvdcn {
                 const at::Tensor &input,
                 const at::Tensor &offset,
                 const at::Tensor &mask,
-                const int in_channels,
-                const int height,
-                const int width,
-                const int weight_h,
-                const int weight_w,
-                const int stride_h,
-                const int stride_w,
-                const int pad_h,
-                const int pad_w,
-                const int dilation_h,
-                const int dilation_w,
-                const int out_h,
-                const int out_w,
-                const int batch_sz,
-                const int offset_groups,
-                const int mask_groups,
+                const int64_t in_channels,
+                const int64_t height,
+                const int64_t width,
+                const int64_t weight_h,
+                const int64_t weight_w,
+                const int64_t stride_h,
+                const int64_t stride_w,
+                const int64_t pad_h,
+                const int64_t pad_w,
+                const int64_t dilation_h,
+                const int64_t dilation_w,
+                const int64_t out_h,
+                const int64_t out_w,
+                const int64_t batch_sz,
+                const int64_t offset_groups,
+                const int64_t mask_groups,
                 const bool deformable,
                 const bool modulated,
                 at::Tensor &columns) {
             const int64_t n_kernels = (int64_t) batch_sz * in_channels * out_h * out_w;
-            const int c_per_offset_group = deformable ? in_channels / offset_groups : 1;
-            const int c_per_mask_group = modulated ? in_channels / mask_groups : 1;
+            const int64_t c_per_offset_group = deformable ? in_channels / offset_groups : 1;
+            const int64_t c_per_mask_group = modulated ? in_channels / mask_groups : 1;
 
             AT_DISPATCH_FLOATING_TYPES_AND_HALF(
                     input.scalar_type(), "im2col_cpu", ([&] {
@@ -299,28 +299,28 @@ namespace tvdcn {
                 const at::Tensor &columns,
                 const at::Tensor &offset,
                 const at::Tensor &mask,
-                const int in_channels,
-                const int height,
-                const int width,
-                const int weight_h,
-                const int weight_w,
-                const int stride_h,
-                const int stride_w,
-                const int pad_h,
-                const int pad_w,
-                const int dilation_h,
-                const int dilation_w,
-                const int out_h,
-                const int out_w,
-                const int batch_sz,
-                const int offset_groups,
-                const int mask_groups,
+                const int64_t in_channels,
+                const int64_t height,
+                const int64_t width,
+                const int64_t weight_h,
+                const int64_t weight_w,
+                const int64_t stride_h,
+                const int64_t stride_w,
+                const int64_t pad_h,
+                const int64_t pad_w,
+                const int64_t dilation_h,
+                const int64_t dilation_w,
+                const int64_t out_h,
+                const int64_t out_w,
+                const int64_t batch_sz,
+                const int64_t offset_groups,
+                const int64_t mask_groups,
                 const bool deformable,
                 const bool modulated,
                 at::Tensor &grad_input) {
             const int64_t n_kernels = (int64_t) batch_sz * in_channels * out_h * out_w * weight_h * weight_w;
-            const int c_per_offset_group = deformable ? in_channels / offset_groups : 1;
-            const int c_per_mask_group = modulated ? in_channels / mask_groups : 1;
+            const int64_t c_per_offset_group = deformable ? in_channels / offset_groups : 1;
+            const int64_t c_per_mask_group = modulated ? in_channels / mask_groups : 1;
 
             AT_DISPATCH_FLOATING_TYPES_AND_HALF(
                     columns.scalar_type(), "col2im_cpu", ([&] {
@@ -417,29 +417,29 @@ namespace tvdcn {
                 const at::Tensor &input,
                 const at::Tensor &offset,
                 const at::Tensor &mask,
-                const int in_channels,
-                const int height,
-                const int width,
-                const int weight_h,
-                const int weight_w,
-                const int stride_h,
-                const int stride_w,
-                const int pad_h,
-                const int pad_w,
-                const int dilation_h,
-                const int dilation_w,
-                const int out_h,
-                const int out_w,
-                const int batch_sz,
-                const int offset_groups,
-                const int mask_groups,
+                const int64_t in_channels,
+                const int64_t height,
+                const int64_t width,
+                const int64_t weight_h,
+                const int64_t weight_w,
+                const int64_t stride_h,
+                const int64_t stride_w,
+                const int64_t pad_h,
+                const int64_t pad_w,
+                const int64_t dilation_h,
+                const int64_t dilation_w,
+                const int64_t out_h,
+                const int64_t out_w,
+                const int64_t batch_sz,
+                const int64_t offset_groups,
+                const int64_t mask_groups,
                 const bool deformable,
                 const bool modulated,
                 at::Tensor &grad_offset) {
             if (!deformable) return;
             const int64_t n_kernels = (int64_t) batch_sz * offset_groups * out_h * out_w * weight_h * weight_w * 2;
-            const int c_per_offset_group = deformable ? in_channels / offset_groups : 1;
-            const int c_per_mask_group = modulated ? in_channels / mask_groups : 1;
+            const int64_t c_per_offset_group = deformable ? in_channels / offset_groups : 1;
+            const int64_t c_per_mask_group = modulated ? in_channels / mask_groups : 1;
 
             AT_DISPATCH_FLOATING_TYPES_AND_HALF(
                     columns.scalar_type(), "deform_conv2d_compute_grad_offset_cpu", ([&] {
@@ -533,29 +533,29 @@ namespace tvdcn {
                 const at::Tensor &columns,
                 const at::Tensor &input,
                 const at::Tensor &offset,
-                const int in_channels,
-                const int height,
-                const int width,
-                const int weight_h,
-                const int weight_w,
-                const int stride_h,
-                const int stride_w,
-                const int pad_h,
-                const int pad_w,
-                const int dilation_h,
-                const int dilation_w,
-                const int out_h,
-                const int out_w,
-                const int batch_sz,
-                const int offset_groups,
-                const int mask_groups,
+                const int64_t in_channels,
+                const int64_t height,
+                const int64_t width,
+                const int64_t weight_h,
+                const int64_t weight_w,
+                const int64_t stride_h,
+                const int64_t stride_w,
+                const int64_t pad_h,
+                const int64_t pad_w,
+                const int64_t dilation_h,
+                const int64_t dilation_w,
+                const int64_t out_h,
+                const int64_t out_w,
+                const int64_t batch_sz,
+                const int64_t offset_groups,
+                const int64_t mask_groups,
                 const bool deformable,
                 const bool modulated,
                 at::Tensor &grad_mask) {
             if (!modulated) return;
             const int64_t n_kernels = (int64_t) batch_sz * mask_groups * out_h * out_w * weight_h * weight_w;
-            const int c_per_offset_group = deformable ? in_channels / offset_groups : 1;
-            const int c_per_mask_group = modulated ? in_channels / mask_groups : 1;
+            const int64_t c_per_offset_group = deformable ? in_channels / offset_groups : 1;
+            const int64_t c_per_mask_group = modulated ? in_channels / mask_groups : 1;
 
             AT_DISPATCH_FLOATING_TYPES_AND_HALF(
                     columns.scalar_type(), "deform_conv2d_compute_grad_mask_cpu", ([&] {
